+++ pyo3-macros-backend/src/defs.rs
@@ -420,6 +420,12 @@ pub const NUM: Proto = Proto {
         MethodProto::new("__imod__", "PyNumberIModProtocol")
             .args(&["Other"])
             .has_self(),
+        // See https://bugs.python.org/issue36379
+        #[cfg(Py_3_8)]
+        MethodProto::new("__ipow__", "PyNumberIPowProtocol")
+            .args(&["Other", "Modulo"])
+            .has_self(),
+        #[cfg(not(Py_3_8))]
         MethodProto::new("__ipow__", "PyNumberIPowProtocol")
             .args(&["Other"])
             .has_self(),